Added a system to preload all sprites and objects
Added a system to preload all sprites.
Fixed the v-sync issues. However, due   to a problem with sprite conversion, had to create a separate display mode  in order to facilitate sprite preloading. I then had to quit the display and re-initialize it. For some reason this doesn't break the converted sprites.
I also added a grenade and for the first time in my life, properly implemented bouncing physics.